+++ src/main/java/org/co2/kanban/rest/project/task/allocation/AllocationResource.java
@@ -27,11 +27,11 @@ public class AllocationResource extends IdentifiableResourceSupport<Allocation> 
         return this.getBean().getAllocationDate();
     }
 
-    public Float getTimeSpent() {
+    public Double getTimeSpent() {
         return this.getBean().getTimeSpent();
     }
 
-    public Float getTimeRemains() {
+    public Double getTimeRemains() {
         return this.getBean().getTimeRemains();
     }
 }
